Move query, acl, and aggregation packages
