+++ package.json
@@ -11,11 +11,13 @@
     "lodash": "~4.17.4",
     "loglevel": "~1.5.1",
     "qs": "~6.5.1",
-    "rxjs": "~5.5.2"
+    "rxjs": "~5.5.2",
+    "sift": "~5.0.0"
   },
   "devDependencies": {
     "babel-preset-env": "~1.6.1",
     "babel-register": "~6.26.0",
+    "chai": "~4.1.2",
     "expect": "~1.20.2",
     "mocha": "~4.0.1"
   }
