--- packages/kinvey-aggregation/package.json
@@ -1,22 +1,0 @@
-{
-  "private": true,
-  "version": "3.9.3-beta.7",
-  "name": "kinvey-aggregation",
-  "main": "src/index.js",
-  "dependencies": {
-    "kinvey-errors": "^3.9.3-beta.7",
-    "kinvey-query": "^3.9.3-beta.7",
-    "kinvey-utils": "^3.9.3-beta.7",
-    "lodash": "4.17.4"
-  },
-  "devDependencies": {
-    "chai": "4.1.2",
-    "chai-as-promised": "7.1.1",
-    "mocha": "4.0.1",
-    "sinon": "4.0.1",
-    "sinon-chai": "2.14.0"
-  },
-  "scripts": {
-    "test": "mocha -s 250 test/**/*.spec.js"
-  }
-}
